Added first navbar routing Moved the home style, layout and page in the home folder. Added a search layout and page in the new search folder. Edited the old layout and page in the root app folder. Added the routing to the Navbar.jsx component. Changed the global stylesheet extenction from css to scss.
