--- src/app/home.scss
@@ -1,41 +1,0 @@
-
-
-nav {
-    height: auto;
-    width: 100%;
-    position: fixed;
-    bottom: 0;
-    background: var(--neutral-color);
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-
-    a {
-        margin: auto;
-        padding: 10px 20px;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-    }
-}
-
-main {
-    margin-top: 64px;
-    margin-bottom: 62px;
-    padding: 0;
-    display: flex;
-    flex-direction: column;
-    background: var(--neutral-color);
-    overflow-y: hide;
-    #content-panel {
-        margin: 0;
-        padding: 20px;
-        border-radius: 12px;
-        background: var(--secondary-color);
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        overflow-y: scroll;
-    }
-}
-
