+++ src/app/layout.jsx
@@ -1,13 +1,5 @@
 import React from 'react';
-import './globals.css';
-import './home.scss';
-import { Open_Sans } from 'next/font/google';
-import MainPanel from '../components/mainpanel/MainPanel';
-import SimpleBottomNavigation from '../components/navbar/Navbar';
-import SidePanel from '../components/appbar/SidePanel'
-import ButtonCreate from '../components/buttoncreate/ButtonCreate'
-
-const font = Open_Sans({ subsets: ['latin'] });
+import './globals.scss';
 
 export const metadata = {
   title: 'Create Next App',
@@ -21,16 +13,7 @@ export default function RootLayout({ children }) {
         <meta name="viewport" content="width=device-width,initial-scale=1" />
         <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap"/>
       </head>
-      <body className={font.className}>
-        <MainPanel name="Home"/>
-        <div id="central-panel">
-          {children}
-          <ButtonCreate/>
-        </div>
-        {/*navbar Mobile in basso*/}
-        <SimpleBottomNavigation name="NavbarF"/>
-        <SidePanel/>
-      </body>
+      {children}
     </html>
   );
 }
