+++ src/components/navbar/Navbar.jsx
@@ -8,6 +8,8 @@ import SearchIcon from '@mui/icons-material/Search';
 import EventAvailableIcon from '@mui/icons-material/EventAvailable';
 import ChatIcon from '@mui/icons-material/Chat';
 import './navbar.scss';
+import { useRouter } from 'next/navigation'
+import { Link } from '@mui/material';
 
 export function Navbar() {
   const [value, setValue] = React.useState(0);
@@ -16,6 +18,8 @@ export function Navbar() {
   const CustomIcon = ({ nameIcon: NameIcon }) => {
     return <NameIcon className="icon" />;
   };
+
+  const router = useRouter()
 
   return (
     <Box className="stickToBottom" >
@@ -27,10 +31,10 @@ export function Navbar() {
           setValue(newValue);
         }}
       >
-        <BottomNavigationAction label="Home" icon={<CustomIcon nameIcon={HomeIcon} />} />
-        <BottomNavigationAction label="Cerca" icon={<CustomIcon nameIcon={SearchIcon} />} />
-        <BottomNavigationAction label="Attività" icon={<CustomIcon nameIcon={EventAvailableIcon} />} />
-        <BottomNavigationAction label="Chat" icon={<CustomIcon nameIcon={ChatIcon} />} />
+        <BottomNavigationAction label="Home" icon={<CustomIcon nameIcon={HomeIcon} />} onClick={() => router.push('/home')} />
+        <BottomNavigationAction label="Cerca" icon={<CustomIcon nameIcon={SearchIcon} />} onClick={() => router.push('/search')} />
+        <BottomNavigationAction label="Attività" icon={<CustomIcon nameIcon={EventAvailableIcon} />} onClick={() => router.push('/activity')} />
+        <BottomNavigationAction label="Chat" icon={<CustomIcon nameIcon={ChatIcon} />} onClick={() => router.push('/chat')} />
       </BottomNavigation>
     </Box>
   );
